Tries to address a few problems the chat view has and simplifies the code (a tiny bit).
just reduce opacity when the view is at the end. Having a scroll flickering doesn't feel good and helps reduce the stress the scrollbar adds in such a small UI when it's not needed. reduce the impact on the bug that appears when the chat is growing too fast

REVIEW: 104554