Add plasma-themeability to the menu on the presence applet
The Presence applet will show a Plasma-themed icon now if
available. However, the context menu items for statuses are still
all Oxygen, so the user experience is inconsistent -- if the
user clicks on an icon for status, they won't get that icon
on the plasmoid. This patch will make it so that those menu items
are also Plasma-themed.

REVIEW: 107545