Fix hovering contact buttons
If we have the icon item as the target for the tooltip, the tool button's
hovering state gets broken and it only works when the mouse is on the
margin between the contact picture and the button frame.
This patch fixes this by attaching to the parent item, so that it keeps
getting the mouse events.

REVIEW: 109038