Chat plasmoid: Polish sizes
I noticed there was a problem where the plasmoid would get the minimum
size at some point and started to investigate. It turned out that there
was an ugly property loop between the minimumSize and the plasmoid's size.
I guess QML is not reporting it because it happens in the Plasma side.

What this patch does is to make sure that we let the plasmoid's width/height be
0 when it's empty.

Also it reads the form factor and in case it's horizontal, it uses the
horizontal flow, despite its size.

REVIEW: 109304