Simplify the behavior of the main UI file in the chat plasmoid
Provide a bigger minimum size, I thought it would never get to the
minimum but it does sometimes, so we can better give a bigger minimum
(which is still too small for a proper usage).
Rename preferredWidth/Height -> implicit*. It seems like it's the
proper way to call these.
When the plasmoid is empty, offer an empty preferred (uh implicit) size.

REVIEW: 109699