Improve integration with hiding panels
This patch changes 2 things:
- Changes the code from a js function to setting a qml property. This is
better because we don't have to connect to all the changing signals to
re-evaluate (there were some weird problems in this regard).
- Also it makes it so when you have a dialog opened, the status is set to
AcceptingInputStatus, this way the panel doesn't hide while you're writing.
This is useful because this way you can see if somebody talks to you while
you're chatting already.

REVIEW: 109947