Use status message from the passed presence
Right now we always take the status message of the current presence and
apply it to the selected presence from the plasmoid. With this it will
always use whatever status message is passed through the KTp::Presence.

Also makes it future-proof (tm) for when we implement custom status
messages for the plasmoid.

Reviewed-by: David Edmundson
BUG: 292305
FIXED-IN: 0.6.1