Chat Plasmoid: Small usability tweaks
Makes sure the text field doesn't have the focus when it's closed (yeah, it
happens, you can write to a non-visible window)
Removes a workaround I introduced thinking would be magic and solve all
problems (because it's hard to reproduce). It didn't.
Makes sure the plasmoid knows there's no currentIndex when we've closed all
windows.

REVIEW: 110412